Fix compile warnings from Mentor Questa
Dozens of warnings like this one:

** Warning: ** while parsing file included at ./src/riscv_instr_pkg.sv(779)
** while parsing macro expansion: 'DV_CHECK_RANDOMIZE_WITH_FATAL' starting at ./src/riscv_page_table_list.sv(166)
** at ./src/riscv_page_table_list.sv(166): (vlog-2643) Unterminated string literal continues onto next line.

The problem was the way the DV_CHECK_RANDOMIZE_WITH_FATAL macro
includes the "with {...}" code in the error message it prints. That
code typically has newlines in it, which results in non-LRM-compliant
multiline string literals.

This fixes that by removing the "with {...}" code from the error message.